+++ packages/terra-table/tests/wdio/single-select-table-spec.js
@@ -1,73 +1,80 @@
-const viewports = Terra.viewports('medium');
-
-describe('Single-Select Table', () => {
+Terra.describeViewports('Single-Select Table', ['medium'], () => {
   describe('Single-Select Table - Interactions', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-table/table/single-select-table/selectable-table');
       browser.waitForVisible('table');
     });
 
-    // reset focusable point for keyboard navigation
-    beforeEach(() => browser.click('[data-terra-table-header-cell]:nth-child(1)'));
-
     describe('Single-Select Table - Click Operations', () => {
       it('selects the second row', () => {
         browser.click('[class*="row"]:nth-child(1)');
       });
 
-      Terra.it.matchesScreenshot('first-row', { viewports });
+      Terra.it.validatesElement('first-row');
 
       it('selects the second row', () => {
         browser.click('[class*="row"]:nth-child(2)');
       });
 
-      Terra.it.matchesScreenshot('second-row', { viewports });
+      Terra.it.matchesScreenshot('second-row');
 
       it('selects the third row', () => {
         browser.click('[class*="row"]:nth-child(3)');
       });
 
-      Terra.it.matchesScreenshot('third-row', { viewports });
+      Terra.it.matchesScreenshot('third-row');
     });
 
     describe('Single-Select Table - Keyboard ENTER key', () => {
+      // reset focusable point for keyboard navigation
+      before(() => {
+        browser.click('[data-terra-table-header-cell]:nth-child(1)');
+        browser.pause(50);
+      });
+
       it('selects the first row', () => {
         browser.keys(['Tab', 'Enter']);
       });
 
-      Terra.it.matchesScreenshot('first-row', { viewports });
+      Terra.it.validatesElement('first-row');
 
       it('selects the second row', () => {
         browser.keys(['Tab', 'Enter']);
       });
 
-      Terra.it.matchesScreenshot('second-row', { viewports });
+      Terra.it.matchesScreenshot('second-row');
 
       it('selects the third row', () => {
         browser.keys(['Tab', 'Enter']);
       });
 
-      Terra.it.matchesScreenshot('third-row', { viewports });
+      Terra.it.matchesScreenshot('third-row');
     });
 
     describe('Single-Select Table - Keyboard SPACE key', () => {
+      // reset focusable point for keyboard navigation
+      before(() => {
+        browser.click('[data-terra-table-header-cell]:nth-child(1)');
+        browser.pause(50);
+      });
+
       it('selects the first row', () => {
         browser.keys(['Tab', 'Space']);
       });
 
-      Terra.it.matchesScreenshot('first-row', { viewports });
+      Terra.it.validatesElement('first-row');
 
       it('selects the second row', () => {
         browser.keys(['Tab', 'Space']);
       });
 
-      Terra.it.matchesScreenshot('second-row', { viewports });
+      Terra.it.matchesScreenshot('second-row');
 
       it('selects the third row', () => {
         browser.keys(['Tab', 'Space']);
       });
 
-      Terra.it.matchesScreenshot('third-row', { viewports });
+      Terra.it.matchesScreenshot('third-row');
     });
   });
 
@@ -81,30 +88,30 @@ describe('Single-Select Table', () => {
       browser.click('[class*="row"]:nth-child(1)');
     });
 
-    Terra.it.matchesScreenshot('first-row', { viewports });
+    Terra.it.matchesScreenshot('first-row');
 
     it('selects the second row', () => {
       browser.click('[class*="row"]:nth-child(2)');
     });
 
-    Terra.it.matchesScreenshot('second-row', { viewports });
+    Terra.it.matchesScreenshot('second-row');
 
     it('selects the third row', () => {
       browser.click('[class*="row"]:nth-child(3)');
     });
 
-    Terra.it.matchesScreenshot('third-row', { viewports });
+    Terra.it.matchesScreenshot('third-row');
   });
 
   describe('Single-Select Table - Preselected', () => {
     before(() => browser.url('/#/raw/tests/terra-table/table/single-select-table/single-row-selectable-table-preselected'));
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.validatesElement();
   });
 
   describe('Single-Select Table - No Rows', () => {
     before(() => browser.url('/#/raw/tests/terra-table/table/single-select-table/single-row-selectable-table-non-selectable'));
 
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.validatesElement();
   });
 });